move astroport cl utils to utils
