+++ e2e/src/utils/mod.rs
@@ -1,3 +1,4 @@
+pub mod astroport;
 pub mod authorization;
 pub mod base_account;
 pub mod ethereum;
